+++ acceptance-tests/src/test/java/tech/pegasys/pantheon/tests/acceptance/dsl/node/PantheonNode.java
@@ -14,6 +14,33 @@ package tech.pegasys.pantheon.tests.acceptance.dsl.node;
 
 import static org.apache.logging.log4j.LogManager.getLogger;
 
+import com.google.common.base.MoreObjects;
+import com.google.common.io.MoreFiles;
+import com.google.common.io.RecursiveDeleteOption;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.net.ConnectException;
+import java.net.URI;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+import java.util.Properties;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import org.apache.logging.log4j.Logger;
+import org.awaitility.Awaitility;
+import org.awaitility.core.ConditionTimeoutException;
+import org.java_websocket.exceptions.WebsocketNotConnectedException;
+import org.web3j.protocol.Web3jService;
+import org.web3j.protocol.core.JsonRpc2_0Web3j;
+import org.web3j.protocol.http.HttpService;
+import org.web3j.protocol.websocket.WebSocketClient;
+import org.web3j.protocol.websocket.WebSocketListener;
+import org.web3j.protocol.websocket.WebSocketService;
+import org.web3j.utils.Async;
 import tech.pegasys.pantheon.cli.EthNetworkConfig;
 import tech.pegasys.pantheon.controller.KeyPairUtil;
 import tech.pegasys.pantheon.crypto.SECP256K1.KeyPair;
@@ -32,35 +59,6 @@ import tech.pegasys.pantheon.tests.acceptance.dsl.transaction.PantheonWeb3j;
 import tech.pegasys.pantheon.tests.acceptance.dsl.transaction.PermissioningJsonRpcRequestFactory;
 import tech.pegasys.pantheon.tests.acceptance.dsl.transaction.Transaction;
 import tech.pegasys.pantheon.tests.acceptance.dsl.waitcondition.WaitCondition;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.net.ConnectException;
-import java.net.URI;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.Properties;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
-
-import com.google.common.base.MoreObjects;
-import com.google.common.io.MoreFiles;
-import com.google.common.io.RecursiveDeleteOption;
-import org.apache.logging.log4j.Logger;
-import org.awaitility.Awaitility;
-import org.awaitility.core.ConditionTimeoutException;
-import org.java_websocket.exceptions.WebsocketNotConnectedException;
-import org.web3j.protocol.Web3jService;
-import org.web3j.protocol.core.JsonRpc2_0Web3j;
-import org.web3j.protocol.http.HttpService;
-import org.web3j.protocol.websocket.WebSocketClient;
-import org.web3j.protocol.websocket.WebSocketListener;
-import org.web3j.protocol.websocket.WebSocketService;
-import org.web3j.utils.Async;
 
 public class PantheonNode implements Node, NodeConfiguration, RunnableNode, AutoCloseable {
 
@@ -82,7 +80,6 @@ public class PantheonNode implements Node, NodeConfiguration, RunnableNode, Auto
   private final boolean devMode;
 
   private List<String> bootnodes = new ArrayList<>();
-  private PantheonWeb3j pantheonWeb3j;
   private JsonRequestFactories jsonRequestFactories;
   private Optional<EthNetworkConfig> ethNetworkConfig = Optional.empty();
 
@@ -196,11 +193,9 @@ public class PantheonNode implements Node, NodeConfiguration, RunnableNode, Auto
       throw new RuntimeException("Error connection to WebSocket endpoint", e);
     }
 
-    if (pantheonWeb3j != null) {
-      pantheonWeb3j.shutdown();
+    if (jsonRequestFactories != null) {
+      jsonRequestFactories.shutdown();
     }
-
-    pantheonWeb3j = new PantheonWeb3j(webSocketService, 2000, Async.defaultExecutorService());
   }
 
   private void checkIfWebSocketEndpointIsAvailable(final String url) {
