+++ acceptance-tests/src/test/java/tech/pegasys/pantheon/tests/acceptance/dsl/transaction/JsonRequestFactories.java
@@ -51,4 +51,8 @@ public class JsonRequestFactories {
   public PermissioningJsonRpcRequestFactory perm() {
     return perm;
   }
+
+  public void shutdown() {
+    netEth.shutdown();
+  }
 }
